converted config to ansible playbook , added progress bar and made some minor changes
