+++ python/avi/migrationtools/nsxt_converter/nsxt_config_converter.py
@@ -8,6 +8,7 @@ import com.vmware.nsx_policy.infra_client as infra_client
 import com.vmware.nsx_policy.model_client as model_client
 import random
 from com.vmware.vapi.std.errors_client import Error
+from avi.migrationtools.avi_rest_lib import upload_config_to_controller
 
 from avi.migrationtools.nsxt_converter.conversion_util import NsxtConvUtil
 from avi.migrationtools.nsxt_converter.monitor_converter import MonitorConfigConv
@@ -17,11 +18,13 @@ import json
 import avi.migrationtools.nsxt_converter.converter_constants as conv_const
 from avi.migrationtools.nsxt_converter.pools_converter import PoolConfigConv
 from avi.migrationtools.nsxt_converter.profile_converter import ProfileConfigConv
+from avi.migrationtools.nsxt_converter.ssl_profile_converter import SslProfileConfigConv
 
 conv_utils = NsxtConvUtil()
 
 
-def convert(nsx_ip, nsx_un, nsx_pw, nsx_port, output_dir, cloud_name, prefix):
+def convert(nsx_ip, nsx_un, nsx_pw, nsx_port, output_dir, cloud_name, prefix
+            ):
     # load the yaml file attribute in nsxt_attributes.
     nsxt_attributes = conv_const.init("11")
 
@@ -45,6 +48,9 @@ def convert(nsx_ip, nsx_un, nsx_pw, nsx_port, output_dir, cloud_name, prefix):
     profile_converter = ProfileConfigConv(nsxt_attributes)
     profile_converter.convert(alb_config, nsx_lb_config, prefix)
 
+    ssl_profile_converter = SslProfileConfigConv(nsxt_attributes)
+    ssl_profile_converter.convert(alb_config, nsx_lb_config, prefix)
+
     output_path = output_dir + os.path.sep + nsx_ip + os.path.sep + "output"
     if not os.path.exists(output_path):
         os.makedirs(output_path)
@@ -58,3 +64,6 @@ def convert(nsx_ip, nsx_un, nsx_pw, nsx_port, output_dir, cloud_name, prefix):
 
     pp = PrettyPrinter()
     pp.pprint(alb_config)
+
+    return alb_config
+
